1. Added private routes for API's 2. Changed Validations folder structure 3. Minor changes to dropCourses and dropCourses API's

+++ routes/api/courses-api.js
@@ -2,16 +2,19 @@ const express = require("express");
 const router = express.Router();
 const jwt = require("jsonwebtoken");
 const keys = require("../../config/keys");
+const auth = require("../../middleware/auth");
 // Load input validation
-const validateCoursesInput = require("../../validation/course-validation");
-const validateCoursesOutput = require("../../validation/course-validation");
+const validateWriteCoursesInput = require("../../validation/courses-api/writeCourses");
+const validateCoursesOutput = require("../../validation/courses-api/getCourses");
+const validateDropCoursesInput = require("../../validation/courses-api/dropCourses");
+const validateDropAllCoursesInput = require("../../validation/courses-api/dropAllCourses");
 
 // Load courses model
 const Courses = require("../../models/courses");
 
 //get method
-router.post("/getCourses", (req, res) => {
-  const { errors, isValid } = validateCoursesOutput(req.body);
+router.post("/getCourses", auth, (req, res) => {
+  const { errors, isValid } = validateWriteCoursesInput(req.body);
   if (!isValid) {
     return res.status(400).json(errors);
   }
@@ -27,7 +30,7 @@ router.post("/getCourses", (req, res) => {
 });
 
 //write method
-router.post("/writeCourses", (req, res) => {
+router.post("/writeCourses", auth, (req, res) => {
   // Courses Search validation
   const { errors, isValid } = validateCoursesInput(req.body);
   // Check validation
@@ -66,12 +69,17 @@ function GenerateUniqueID() {
 }
 
 //drop courses
-router.post("/dropCourses", (req, res) => {
+router.post("/dropCourses", auth, (req, res) => {
+  const { errors, isValid } = validateDropCoursesInput(req.body);
+  // Check validation
+  if (!isValid) {
+    return res.status(400).json(errors);
+  }
   Courses.deleteOne({
     email: req.body.email,
     course_name: req.body.course_name,
   }).then((courses) => {
-    if (courses) {
+    if (courses.n == 1) {
       res.json({ courses: "Course deleted." });
     } else {
       return res.status(400).json({ courses: "Course not found." });
@@ -80,12 +88,17 @@ router.post("/dropCourses", (req, res) => {
 });
 
 //drop all courses
-router.post("/dropAllCourses", (req, res) => {
+router.post("/dropAllCourses", auth, (req, res) => {
+  const { errors, isValid } = validateDropAllCoursesInput(req.body);
+  if (!isValid) {
+    return res.status(400).json(errors);
+  }
   Courses.deleteMany({
     email: req.body.email,
   }).then((courses) => {
-    if (courses) {
-      res.json({ courses: "Deleted all courses." });
+    if (courses.n >= 1 ) {
+      res.json({ courses: "Deleted all courses.",
+                  num_deletes: courses.n});
     } else {
       return res.status(400).json({ courses: "Course(s) not found." });
     }
